+++ packages/atomic/src/utils/string-utils.ts
@@ -1,0 +1,3 @@
+export function regexEncode(value: string): string {
+  return value.replace(/[-[\]/{}()*+?.\\^$|]/g, '\\$&');
+}