+++ packages/atomic/stencil.config.ts
@@ -59,7 +59,7 @@ export const config: Config = {
     transformIgnorePatterns: [],
   },
   devServer: {
-    reloadStrategy: 'hmr',
+    reloadStrategy: 'pageReload',
   },
   plugins: [
     inlineSvg(),