+++ packages/quantic/force-app/main/default/lwc/testUtils/coveoHeadlessStub.js
@@ -1,0 +1,15 @@
+export class MockEngine {
+  static getSampleConfiguration() {
+    return {}
+  }
+}
+
+export const CoveoHeadlessStub = {
+  HeadlessEngine: MockEngine,
+  SearchActions: {
+    executeSearch: () => {}
+  },
+  AnalyticsActions: {
+    logInterfaceLoad: () => {}
+  }
+}